type of product_id and cart_id changed to bigint

+++ db/schema.rb
@@ -18,8 +18,8 @@ ActiveRecord::Schema.define(version: 2020_02_23_080004) do
   end
 
   create_table "line_items", force: :cascade do |t|
-    t.integer "product_id", null: false
-    t.integer "cart_id"
+    t.bigint "product_id", null: false
+    t.bigint "cart_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "quantity", default: 1
